+++ fidget-core/src/compiler/reg_tape.rs
