+++ fidget-core/src/eval/bulk.rs
