+++ fidget-core/src/types/interval.rs
