+++ fidget-gui/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "fidget-gui"
+description = "GUI abstractions for Fidget"
+readme = "../README.md"
+
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+fidget-core = { path = "../fidget-core", version = "=0.4.0" }
+
+nalgebra.workspace = true
+serde.workspace = true
