+++ fidget-raster/src/config.rs
@@ -207,7 +207,7 @@ impl<const N: usize> Tile<N> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use fidget_core::{gui::View2, render::ImageSize};
+    use fidget_core::render::ImageSize;
 
     #[test]
     fn test_default_render_config() {
@@ -251,55 +251,6 @@ mod test {
                 config.image_size.height() as f32 - 1.0,
             )),
             Point2::new(1.0, -1.0)
-        );
-    }
-
-    #[test]
-    fn test_camera_render_config() {
-        let config = ImageRenderConfig {
-            image_size: ImageSize::from(512),
-            world_to_model: View2::from_center_and_scale(
-                nalgebra::Vector2::new(0.5, 0.5),
-                0.5,
-            )
-            .world_to_model(),
-            ..Default::default()
-        };
-        let mat = config.mat();
-        assert_eq!(
-            mat.transform_point(&Point2::new(0.0, -1.0)),
-            Point2::new(0.0, 1.0)
-        );
-        assert_eq!(
-            mat.transform_point(&Point2::new(512.0, -1.0)),
-            Point2::new(1.0, 1.0)
-        );
-        assert_eq!(
-            mat.transform_point(&Point2::new(512.0, 511.0)),
-            Point2::new(1.0, 0.0)
-        );
-
-        let config = ImageRenderConfig {
-            image_size: ImageSize::from(512),
-            world_to_model: View2::from_center_and_scale(
-                nalgebra::Vector2::new(0.5, 0.5),
-                0.25,
-            )
-            .world_to_model(),
-            ..Default::default()
-        };
-        let mat = config.mat();
-        assert_eq!(
-            mat.transform_point(&Point2::new(0.0, -1.0)),
-            Point2::new(0.25, 0.75)
-        );
-        assert_eq!(
-            mat.transform_point(&Point2::new(512.0, -1.0)),
-            Point2::new(0.75, 0.75)
-        );
-        assert_eq!(
-            mat.transform_point(&Point2::new(512.0, 511.0)),
-            Point2::new(0.75, 0.25)
         );
     }
 }
